django_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==23.3
Pillow==10.1.0
stripe==7.8.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-storages==1.14.2
boto3==1.34.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
gunicorn==21.2.0
whitenoise==6.6.0
channels==4.0.0
channels-redis==4.1.0
django-redis==5.4.0
elasticsearch-dsl==8.11.0
openai==1.3.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2